requirements.txt
fastapi
uvicorn[standard]
feedparser
vaderSentiment
python-jose
